Separate vscode plugin into multiple files
